Address dazuma's comments
