+++ Gemfile
@@ -1,6 +1,5 @@
 source "https://rubygems.org"
 
 gem "concurrent-ruby", require: "concurrent"
-# uncomment and run `bundle install` before running tests
-# gem "gemstash", git: "https://github.com/bundler/gemstash.git"
+gem "gemstash", git: "https://github.com/bundler/gemstash.git", ref: "a5a78e2"
 gemspec
