+++ test/google/cloud/gemserver/cli/request_test.rb
@@ -27,7 +27,7 @@ describe Google::Cloud::Gemserver::CLI::Request do
     it "calls send_req with the correct arguments" do
       bkd = GCG::CLI::Request.new "google.com"
       mock = Minitest::Mock.new
-      mock.expect :call, nil, ["post", "/api/v1/key", {permissions: nil}]
+      mock.expect :call, nil, [Net::HTTP::Post, "/api/v1/key", {permissions: nil}]
       bkd.stub :send_req, mock do
         bkd.create_key
         mock.verify
@@ -39,7 +39,7 @@ describe Google::Cloud::Gemserver::CLI::Request do
     it "calls send_req with the correct arguments" do
       bkd = GCG::CLI::Request.new "google.com"
       mock = Minitest::Mock.new
-      mock.expect :call, nil, ["put", "/api/v1/key", {key: "key"}]
+      mock.expect :call, nil, [Net::HTTP::Put, "/api/v1/key", {key: "key"}]
       bkd.stub :send_req, mock do
         bkd.delete_key "key"
         mock.verify
@@ -51,7 +51,7 @@ describe Google::Cloud::Gemserver::CLI::Request do
     it "calls send_req with the correct arguments" do
       bkd = GCG::CLI::Request.new "google.com"
       mock = Minitest::Mock.new
-      mock.expect :call, nil, ["get", "/api/v1/stats"]
+      mock.expect :call, nil, [Net::HTTP::Get, "/api/v1/stats"]
       bkd.stub :send_req, mock do
         bkd.stats
         mock.verify
